♻️ Refactor related to linkedin asset usage
