+++ src/components/advocate-title.tsx
@@ -13,7 +13,13 @@ export default function AdvocateTitle({
     <div className="flex items-center justify-between mt-3">
       <span className="text-lg">{title}</span>
       <a href={url} className="relative h-full w-auto" target="_blank">
-        <Image src={icon} alt={`${title} icon`} width={22} height={22} />
+        <Image
+          src={icon}
+          alt={`${title} icon`}
+          width={22}
+          height={22}
+          className="grayscale hover:grayscale-0 transition-all duration-200 delay-50"
+        />
       </a>
     </div>
   );
